[TE-5] Renamed files
